Random: deprecate randjump in favor of Future.randjump and accumulate
